+++ src/DotnetAgentHarness.Cli.Tests/Services/TransactionManagerTests.cs
@@ -102,17 +102,16 @@ public class TransactionManagerTests : IDisposable
 
     protected virtual void Dispose(bool disposing)
     {
-        if (!this.disposedValue)
+        if (this.disposedValue)
         {
-            if (disposing)
-            {
-                if (Directory.Exists(this.testDir))
-                {
-                    Directory.Delete(this.testDir, true);
-                }
-            }
-
-            this.disposedValue = true;
+            return;
         }
+
+        if (disposing && Directory.Exists(this.testDir))
+        {
+            Directory.Delete(this.testDir, true);
+        }
+
+        this.disposedValue = true;
     }
 }
